requirements.txt
# --- Core ---
numpy>=1.23
tqdm>=4.65
PyYAML>=6.0

# --- Embeddings / LLM ---
sentence-transformers>=2.2.2
transformers>=4.40.0
torch>=2.0.1

# --- FAISS ---
faiss-cpu>=1.7.4   # Nếu có GPU: faiss-gpu>=1.7.4

# --- LangChain / Text chunking ---
langchain>=0.1.0
tiktoken>=0.4.0

# --- Streamlit App ---
streamlit>=1.26.0

# --- API ---
fastapi>=0.102.0
uvicorn>=0.23.2
pydantic>=2.5.1

# --- Data processing / Visualization ---
pandas>=2.1.0
matplotlib>=3.8.0
seaborn>=0.12.2
scikit-learn>=1.3.0

# --- PDF / Text parsing ---
PyPDF2>=3.0.1
beautifulsoup4>=4.12.2

# --- Optional logging ---
loguru>=0.7.0
